feat: Wave 219 — 6 test files (362 tests) across IR validation, runtime2 parser, tool codegen, GLR first-follow, common filters, error recovery

+++ runtime2/tests/tree_cursor_v3_comprehensive.rs
@@ -35,9 +35,9 @@ fn chain_tree(depth: usize) -> Tree {
     if depth == 0 {
         return leaf(0, 0, 1);
     }
-    let mut current = leaf(depth as u32, 0, (depth + 1) as usize);
+    let mut current = leaf(depth as u32, 0, depth + 1);
     for i in (0..depth).rev() {
-        current = Tree::new_for_testing(i as u32, 0, (depth + 1) as usize, vec![current]);
+        current = Tree::new_for_testing(i as u32, 0, depth + 1, vec![current]);
     }
     current
 }
